Clean up vector classes
We had essentially 3 classes to manage an x,y,z with various helper
functions. I merged two of them and left the point class for now. I also
fixed the unit tests so they will run and a lot of various problems like
header guards and what not.

The structure has been rearranged so code that needs to be tested (non
gui code) is moved into the common library.